Added a method to print the game to the console
Added a method to the Game class that populates the grid, then prints the initial state; then evolves the grid, and continues to print it; this occurs for a set number of times stated by the user at the start of the program. Also created a Main class that's used to take user input, creates a new game, and then prints the grid.