Populated the grid, and added a period to the timer
Added a period to the timer, so that the actions that the timer undertakes are repeated (also set the long delay to 0), and initialised the timer itself in the Controllers constructor. Also populated the grid in the GridDisplay as it was initially empty, resulting in an empty grid being displayed.